# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Notewise AI Backend","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/study-sets":{"get":{"summary":"Get Study Sets","operationId":"get_study_sets_api_study_sets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/reviews/today":{"get":{"summary":"Get Reviews Today","operationId":"get_reviews_today_api_reviews_today_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/generate":{"post":{"summary":"Generate","description":"Accepts a payload from the frontend (which includes file upload handling).\nFor simplicity this endpoint expects the frontend to have already uploaded the PDF\nbytes (base64 or multipart). Here we assume 'pdf_bytes' key may contain raw bytes.\nAdapt to how your frontend sends the file.","operationId":"generate_api_generate_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/study-set/{set_id}/flashcards":{"get":{"summary":"Get Flashcards","operationId":"get_flashcards_api_study_set__set_id__flashcards_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"set_id","in":"path","required":true,"schema":{"type":"integer","title":"Set Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/quiz/{set_id}":{"get":{"summary":"Get Quiz By Set","operationId":"get_quiz_by_set_api_quiz__set_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"set_id","in":"path","required":true,"schema":{"type":"integer","title":"Set Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/quiz/complete":{"post":{"summary":"Quiz Complete","operationId":"quiz_complete_api_quiz_complete_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizCompletePayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/arena/{set_id}":{"get":{"summary":"Get Arena Challenge By Set","operationId":"get_arena_challenge_by_set_api_arena__set_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"set_id","in":"path","required":true,"schema":{"type":"integer","title":"Set Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/arena/session/start":{"post":{"summary":"Start Arena Session","operationId":"start_arena_session_api_arena_session_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartArenaSessionPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/arena/session/{session_id}":{"get":{"summary":"Get Arena Session","operationId":"get_arena_session_api_arena_session__session_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QuizCompletePayload":{"properties":{"set_id":{"type":"integer","title":"Set Id"},"session_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Session Id"},"answers":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Answers"}},"type":"object","required":["set_id","answers"],"title":"QuizCompletePayload"},"StartArenaSessionPayload":{"properties":{"set_id":{"type":"integer","title":"Set Id"},"num_questions":{"type":"integer","title":"Num Questions","default":1}},"type":"object","required":["set_id"],"title":"StartArenaSessionPayload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/login"}}}}}}